+++ spacepackets/cfdp/pdu/helper.py
@@ -1,20 +1,24 @@
 from __future__ import annotations
-from typing import cast, Union, Type, Optional
+
+from typing import Optional, Type, Union, cast, Any
+
+import deprecation
 
 from spacepackets.cfdp import PduType
 from spacepackets.cfdp.pdu import (
-    MetadataPdu,
     AbstractFileDirectiveBase,
-    DirectiveType,
     AckPdu,
-    NakPdu,
-    FinishedPdu,
+    DirectiveType,
     EofPdu,
+    FinishedPdu,
     KeepAlivePdu,
+    MetadataPdu,
+    NakPdu,
     PromptPdu,
 )
 from spacepackets.cfdp.pdu.file_data import FileDataPdu
 from spacepackets.cfdp.pdu.header import AbstractPduBase
+from spacepackets.version import get_version
 
 GenericPduPacket = Union[AbstractFileDirectiveBase, AbstractPduBase]
 
@@ -22,8 +26,8 @@ GenericPduPacket = Union[AbstractFileDirectiveBase, AbstractPduBase]
 class PduHolder:
     """Helper type to store arbitrary PDU types and cast them to a concrete PDU type conveniently"""
 
-    def __init__(self, base: Optional[GenericPduPacket]):
-        self.base = base
+    def __init__(self, pdu: Optional[GenericPduPacket]):
+        self.pdu = pdu
 
     def pack(self) -> bytearray:
         if self.base is None:
@@ -31,14 +35,33 @@ class PduHolder:
         return self.base.pack()
 
     @property
+    @deprecation.deprecated(
+        deprecated_in="0.19.0",
+        current_version=get_version(),
+        details="use packet member instead",
+    )
+    def base(self):
+        return self.pdu
+
+    @base.setter
+    @deprecation.deprecated(
+        deprecated_in="0.19.0",
+        current_version=get_version(),
+        details="use packet member instead",
+    )
+    def base(self, base: GenericPduPacket):
+        self.pdu = base
+
+    @property
     def packet_len(self) -> int:
-        if self.base is None:
+        if self.pdu is None:
             return 0
-        return self.base.packet_len
+        return self.pdu.packet_len
 
     @property
     def pdu_type(self) -> PduType:
-        return self.base.pdu_header.pdu_type
+        assert self.pdu is not None
+        return self.pdu.pdu_type
 
     @property
     def is_file_directive(self):
@@ -51,35 +74,34 @@ class PduHolder:
         """
         if not self.is_file_directive:
             return None
-        directive_base = cast(AbstractFileDirectiveBase, self.base)
+        directive_base = cast(AbstractFileDirectiveBase, self.pdu)
         return directive_base.directive_type
 
     def __repr__(self):
-        return f"{self.__class__.__name__}(base={self.base!r}"
+        return f"{self.__class__.__name__}(base={self.pdu!r}"
 
-    def _raise_not_target_exception(self, pdu_type: Type[any]):
-        raise TypeError(f"Stored PDU is not {pdu_type.__name__!r}: {self.base!r}")
+    def _raise_not_target_exception(self, pdu_type: Type[Any]):
+        raise TypeError(f"Stored PDU is not {pdu_type.__name__!r}: {self.pdu!r}")
 
     def _cast_to_concrete_file_directive(
-        self, pdu_type: Type[any], dir_type: DirectiveType
-    ):
+        self, pdu_type: Type[Any], dir_type: DirectiveType
+    ) -> Any:
         if (
             isinstance(self.base, AbstractFileDirectiveBase)
-            and self.base.pdu_type == PduType.FILE_DIRECTIVE
+            and self.pdu.pdu_type == PduType.FILE_DIRECTIVE  # type: ignore
         ):
-            pdu_base = cast(AbstractFileDirectiveBase, self.base)
+            pdu_base = cast(AbstractFileDirectiveBase, self.pdu)
             if pdu_base.directive_type == dir_type:
-                return cast(pdu_type, self.base)
+                return cast(pdu_type, self.pdu)
         self._raise_not_target_exception(pdu_type)
 
-    def to_file_data_pdu(self) -> FileDataPdu:
+    def to_file_data_pdu(self) -> FileDataPdu:  # type: ignore
         if (
-            isinstance(self.base, AbstractPduBase)
-            and self.base.pdu_type == PduType.FILE_DATA
+            isinstance(self.pdu, AbstractPduBase)
+            and self.pdu.pdu_type == PduType.FILE_DATA
         ):
-            return cast(FileDataPdu, self.base)
-        else:
-            self._raise_not_target_exception(FileDataPdu)
+            return cast(FileDataPdu, self.pdu)
+        self._raise_not_target_exception(FileDataPdu)
 
     def to_metadata_pdu(self) -> MetadataPdu:
         return self._cast_to_concrete_file_directive(