+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Moved `pdu_header` abstract property from `AbstractFileDirectiveBase` to `AbstractPduBase`
   class
+- Renamed `TlvHolder` field `base` to `tlv` and `PduHolder` field `base` to `pdu`.
 
 # [v0.18.0] 2023-09-08